googlebigquery_blueprints/requirements.txt
google-cloud-bigquery<3.0.0
pandas==1.3.5
db-dtypes==1.0.4
google-cloud-storage==2.7.0
bq-iterate